Renaming perftest packages
